fix: implementing required changes
