+++ src/utils/custom-error.ts
@@ -1,0 +1,11 @@
+export interface GraphQLServerError {
+  graphQLErrors: GraphQLError[];
+  name: string;
+  message: string;
+}
+
+interface GraphQLError {
+  message: string;
+  code: number;
+  name: string;
+}
