+++ src/utils/storage.ts
@@ -1,0 +1,9 @@
+import AsyncStorage from '@react-native-async-storage/async-storage';
+
+export const asyncStorage = (key: string, value: string) => {
+  try {
+    AsyncStorage.setItem(key, value);
+  } catch {
+    throw 'Chave inexistente';
+  }
+};
